// src/modules/deck/deck.enum.ts
export enum Visibility {
	PUBLIC = "public",
	PROTECTED = "protected",
	PRIVATE = "private",
}

export enum CardStatus {
	KNOWN = "known",
	LEARNING = "learning",
	NEW = "new",
}

export enum DeckOrderBy {
	CREATED_AT = "createdAt",
	UPDATED_AT = "updatedAt",
	OPENED_AT = "openedAt",
	NAME = "name",
}